backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
google-adk>=1.0.0
google-genai
google-cloud-aiplatform
cloud-sql-python-connector[asyncpg]
pydantic>=2.0
alembic
